+++ src/main/java/kr/co/mz/sns/controller/user/UserDetailController.java
@@ -1,8 +1,6 @@
 package kr.co.mz.sns.controller.user;
 
 import jakarta.validation.constraints.NotNull;
-import java.util.List;
-import kr.co.mz.sns.dto.user.InsertUserProfileDtos;
 import kr.co.mz.sns.dto.user.FindUserDetailDto;
 import kr.co.mz.sns.dto.user.GenericUserDetailFileDto;
 import kr.co.mz.sns.dto.user.InsertUserDetailDto;
@@ -63,38 +61,37 @@ public class UserDetailController {
 //        }
 //    }
 
-  @PostMapping
-  public ResponseEntity<InsertUserDetailDto> insert(@RequestBody InsertUserDetailDto insertUserDetailDto) {
-    return ResponseEntity
-        .status(HttpStatus.CREATED)
-        .body(
-            userDetailService.insert(insertUserDetailDto)
+    @PostMapping
+    public ResponseEntity<InsertUserDetailDto> insert(@RequestBody InsertUserDetailDto insertUserDetailDto) {
+        return ResponseEntity
+            .status(HttpStatus.CREATED)
+            .body(
+                userDetailService.insert(insertUserDetailDto)
 
             );
     }
 
-  @PostMapping("/upload")
-  public ResponseEntity<List<GenericUserDetailFileDto>> uploadFile(
-      @RequestParam InsertUserProfileDtos insertUserProfileDtos
-  ){
+//  @PostMapping("/upload")
+//  public ResponseEntity<List<GenericUserDetailFileDto>> uploadFile(
+//      @RequestParam InsertUserProfileDtos insertUserProfileDtos
+//  ){
+//
+//    var fileDtos =
+//  }
 
-    var fileDtos =
-  }
-
-  @GetMapping("/download")
-  public ResponseEntity<InputStreamResource> downloadFile(@RequestParam("fileName") String fileName) {
-    var inputStream = FileStorageService.downloadFile(
-        fileService.findByName(new GenericUserDetailFileDto(fileName))
-    );
-    var headers = new HttpHeaders();
-    headers.add(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + fileName);
-    return ResponseEntity
-        .ok()
-        .headers(headers)
-        .contentType(MediaType.APPLICATION_OCTET_STREAM)
-        .body(new InputStreamResource(inputStream));
-  }
-
+    @GetMapping("/download")
+    public ResponseEntity<InputStreamResource> downloadFile(@RequestParam("fileName") String fileName) {
+        var inputStream = FileStorageService.downloadFile(
+            fileService.findByName(new GenericUserDetailFileDto(fileName))
+        );
+        var headers = new HttpHeaders();
+        headers.add(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + fileName);
+        return ResponseEntity
+            .ok()
+            .headers(headers)
+            .contentType(MediaType.APPLICATION_OCTET_STREAM)
+            .body(new InputStreamResource(inputStream));
+    }
 
 
     @PutMapping
