+++ src/main/java/kr/co/mz/sns/service/comment/CommentService.java
@@ -6,7 +6,6 @@ import kr.co.mz.sns.entity.comment.CommentEntity;
 import kr.co.mz.sns.exception.NotFoundException;
 import kr.co.mz.sns.repository.comment.CommentLikeRepository;
 import kr.co.mz.sns.repository.comment.CommentRepository;
-import kr.co.mz.sns.service.post.PostService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
@@ -16,41 +15,41 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class CommentService {
 
-  private final CommentRepository commentRepository;
-  private final CommentLikeRepository commentLikeRepository;
-  private final PostService postService;
-  private final ModelMapper modelMapper;
+    private final CommentRepository commentRepository;
+    private final CommentLikeRepository commentLikeRepository;
+    //  private final PostService postService;
+    private final ModelMapper modelMapper;
 
 
-  @Transactional
-  public CommentDto insert(CommentDto commentDto) {
-    if (commentDto.isLike()) {
-      postService.like(commentDto.getPostSeq());
+    @Transactional
+    public CommentDto insert(CommentDto commentDto) {
+//        if (commentDto.isLike()) {
+//            postService.like(commentDto.getPostSeq());
+//        }
+
+        CommentEntity commentEntity = commentRepository.save(modelMapper.map(commentDto, CommentEntity.class));
+        return modelMapper.map(commentEntity, CommentDto.class);
     }
 
-    CommentEntity commentEntity = commentRepository.save(modelMapper.map(commentDto, CommentEntity.class));
-    return modelMapper.map(commentEntity, CommentDto.class);
-  }
+    @Transactional
+    public void delete(Long commentSeq) {
+        Optional<CommentEntity> optional = commentRepository.findBySeq(commentSeq);
+        var commentEntity = optional.orElseThrow(() -> new NotFoundException("It is not exist comment"));
 
-  @Transactional
-  public void delete(Long commentSeq) {
-    Optional<CommentEntity> optional = commentRepository.findBySeq(commentSeq);
-    var commentEntity = optional.orElseThrow(() -> new NotFoundException("It is not exist comment"));
-
-    commentRepository.delete(commentEntity);
-  }
-
-  @Transactional
-  public void update(Long commentSeq, CommentDto commentDto) {
-    var optionalComment = commentRepository.findBySeq(commentSeq);
-    var commentEntity = optionalComment.orElseThrow(() -> new NotFoundException("It is not exist comment"));
-
-    if (commentDto.isLike()) {
-      postService.like(commentDto.getPostSeq());
+        commentRepository.delete(commentEntity);
     }
 
-    commentEntity.setContent(commentDto.getContent());
-  }
+    @Transactional
+    public void update(Long commentSeq, CommentDto commentDto) {
+        var optionalComment = commentRepository.findBySeq(commentSeq);
+        var commentEntity = optionalComment.orElseThrow(() -> new NotFoundException("It is not exist comment"));
+
+//        if (commentDto.isLike()) {
+//            postService.like(commentDto.getPostSeq());
+//        }
+
+        commentEntity.setContent(commentDto.getContent());
+    }
 
 //    public CommentLikeDto like(Long commentSeq){
 //        var commentEntityOptional = commentRepository.findBySeq(commentSeq);
